Allow JXL in non-nightly builds
By default, enabling JPEG XL is only possible in nightly builds.
This allows enabling it in regular builds as well (and enables the build
option; hopefully I got all the places to add it?)

It could also be considered to enable `image.jxl.enabled` in the
settings by default as well, so it would "just work".
There are no obvious downsides to that approach that I know of.