Completely remove experimental GenAI feature.
This is a draft intended to spur discussion around the specific technical approach to removing this feature.  Unfortunately, Mozilla didn’t add a compile-time flag to control this, so I went with the maximalist option and reverted *all* code related to this feature; the diff is, consequently, large.  This is likely to have maintenance issues and I don’t expect it to get merged in the current state.  However, I think it’s a good starting point for discussion, and may be useful for some users who wish to compile their own browser.

Most of the code for this feature is in `browser/components/genai`, and the removal of that makes up the bulk of the diff.  I think a simpler approach is to remove `"genai"` from the browser component list in `browser/components/moz.build`, which will exclude that stuff from the build, without needing the large diff to remove it from the source tree.  That said, there are many places in the code which reference that component, and I’m not sure which of them also need to be removed to avoid chrome showing up which won’t work.  I think the `browser/components/preferences/experimental.inc.xhtml` diff likely also needs to be included -- those two changes, plus locking the preferences, seem like a promising route.

I’d welcome any specific technical feedback about *how* to remove the genai feature in this PR, but request that discussions about *whether to* remove it continue in librewolf/issues#1919.